mesa 10.5.4
Upstream removed the Gallium EGL state tracker (st/egl) for 10.4 on the
basis that the same thing could be accomplished with the egl_dri2 driver.
While that is technically true, it required a fair bit of patching to
avoid all use of xcb-dri2 therein for the HW drivers.  This change
obsoleted patches 0006 and 0008, while patch 0001 was upstreamed for 10.5.